refactor(card component): divide card component structure into basic and for discount codes
