+++ src/app/components/card-component.ts
@@ -1,109 +1,35 @@
-import { showErrorMessage, showSucessMessage } from '@utils/toast-messages';
-import { BaseComponent, IAttributes } from './base-component';
-import { ButtonComponent, IButtonAttributes } from './button-component';
-import { IImageAttributes, ImageComponent } from './image-component';
-import { IInputAttributes, InputComponent } from './input-component';
+import { BaseComponent } from "./base-component";
 
 export interface ICardAttributes {
-  imageSrc?: string;
-  title?: string;
-  description?: string;
-  inputValue?: string;
-}
+    imageSrc?: string;
+    title?: string;
+    description?: string;
+    inputValue?: string;
+  }
 
-export class CardComponent extends BaseComponent {
-  constructor(data: Partial<ICardAttributes> = {}, discountContainerAttr?: IAttributes) {
-    const {
-      imageSrc = '',
-      title = '',
-      description = '',
-      inputValue = '',
-    } = data;
+  export class CardComponent extends BaseComponent {
+    protected card: BaseComponent ;
 
-    super({ tag: 'div', classList: ['col', 's12', 'm4', 'center-align'] });
-
-    const card = new BaseComponent({ tag: 'div', classList: ['card'] });
-    this.appendChild(card);
-
-    const cardImage = new BaseComponent({
-      tag: 'div',
-      classList: ['card-image'],
-    });
-    card.appendChild(cardImage);
-
-    if (imageSrc !== '') {
-      const imgAttr: IImageAttributes = {
-        alt: 'img',
-        src: imageSrc,
-      };
-      const img = new ImageComponent(imgAttr);
-      cardImage.appendChild(img);
-    }
-
-    if (title !== '') {
-      const titleSpan = new BaseComponent({
-        tag: 'span',
-        classList: ['card-title'],
-        content: title,
-      });
-      cardImage.appendChild(titleSpan);
-    }
-
-    if (description !== '') {
+    constructor(data: Partial<ICardAttributes> = {}) {
+      const { description = '' } = data;
+  
+      super({ tag: 'div', classList: ['col', 's12', 'm4', 'center-align'] });
+      
+      const card = new BaseComponent({ tag: 'div', classList: ['card'] });
+      this.card = card;
+      this.appendChild(this.card);
+  
       const cardContent = new BaseComponent({
         tag: 'div',
         classList: ['card-content'],
       });
       card.appendChild(cardContent);
-
+  
       const contentParagraph = new BaseComponent({
         tag: 'p',
         content: description,
       });
       cardContent.appendChild(contentParagraph);
     }
-
-    if (discountContainerAttr) {
-      const discountContainer = new BaseComponent(discountContainerAttr);
-      card.appendChild(discountContainer);
-
-      const inputAttr: IInputAttributes = {
-        disabled: true,
-        value: inputValue,
-        classList: ['col s3 red-text offset-s2'],
-        id: 'input',
-      };
-      const input = new InputComponent(inputAttr);
-      discountContainer.appendChild(input);
-
-      const buttonAttr: IButtonAttributes = {
-        content: 'COPY',
-        onClick: () => {
-          const inputElement = document.getElementById(
-            'input'
-          ) as HTMLInputElement;
-          if (inputElement) {
-            navigator.clipboard
-              .writeText(inputElement.value)
-              .then(() => {
-                showSucessMessage('Text copied to clipboard');
-              })
-              .catch(() => {
-                showErrorMessage('Failed to copy text');
-              });
-          }
-        },
-        classList: [
-          'edit-profile-button',
-          'profile-button',
-          'waves-effect',
-          'waves-light',
-          'btn',
-          'no-text-transform',
-        ],
-      };
-      const button = new ButtonComponent(buttonAttr);
-      discountContainer.appendChild(button);
-    }
   }
-}
+
