+++ src/app/views/main/main-page.ts
@@ -2,7 +2,9 @@ import View from '@views/view';
 import { BaseComponent, IAttributes } from '@components/base-component';
 import { IImageAttributes, ImageComponent } from '@components/image-component';
 import carUrl from '@assets/main-image.webp';
-import { CardComponent, ICardAttributes } from '@components/card-component';
+import { ICardAttributes } from '@components/card-component';
+import DiscountCardComponent  from './discount-card';
+
 
 export default class MainPageView extends View {
   private mainContainer!: BaseComponent; // добавлять разные блоки сюда
@@ -38,30 +40,25 @@ export default class MainPageView extends View {
     const promoSection = new BaseComponent(promoSectionAttr);
     this.mainContainer.appendChild(promoSection);
 
-    const discountContainer: IAttributes = {
-      tag: 'div',
-      classList: ['row', 'center-align'],
-    };
-
     const cardAttr1: ICardAttributes = {
       inputValue: 'DISCOUNT10',
       description: 'GET 30% OFF YOUR CART WITH ORDERS OVER 1000$',
     };
-    const card1 = new CardComponent(cardAttr1, discountContainer);
+    const card1 = new DiscountCardComponent(cardAttr1);
     promoSection.appendChild(card1);
 
     const cardAttr2: ICardAttributes = {
       inputValue: 'DISCOUNT20',
-      description: 'TAKE AN EXTRA 20% OFF FOR ORDERS OVER 1000€',
+      description: 'TAKE EXTRA 20% OFF FOR ORDERS OVER 1000€',
     };
-    const card2 = new CardComponent(cardAttr2, discountContainer);
+    const card2 = new DiscountCardComponent(cardAttr2);
     promoSection.appendChild(card2);
 
     const cardAttr3: ICardAttributes = {
       inputValue: 'DISCOUNT30',
       description: 'GET 30% OFF YOUR CART WITH ORDERS OVER 1500$',
     };
-    const card3 = new CardComponent(cardAttr3, discountContainer);
+    const card3 = new DiscountCardComponent(cardAttr3);
     promoSection.appendChild(card3);
   }
 }
